% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_conversion.R
\name{lat_conversion}
\alias{lat_conversion}
\title{Conversion of latitudinal coordinates to grid format}
\usage{
lat_conversion(coord_lat)
}
\arguments{
\item{coord_lon}{latitudinal coordinates}
}
\value{
single numeric value
}
\description{
Takes latitudinal coordinates and uses proper math formula to get grid values
}
\examples{
lat_conversion(20)

}
