% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lon_conversion.R
\name{lon_conversion}
\alias{lon_conversion}
\title{Conversion of longitudinal coordinates to grid format}
\usage{
lon_conversion(coord_lon)
}
\arguments{
\item{coord_lon}{longitudinal coordinates}
}
\value{
single numeric value
}
\description{
Takes longitudinal coordinates and uses proper math formula to get grid values
}
\examples{
lon_conversion(20)

}
