% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_mass_slope.R
\name{regression_mass_slope}
\alias{regression_mass_slope}
\title{Calculation of each species' annual mass slope}
\usage{
regression_mass_slope(mass_df)
}
\arguments{
\item{mass_df}{dataframe containing all annual masses for all species}
}
\value{
dataframe with species and slope columns
}
\description{
Takes annual masses for all species, runs linear regression on all annual
masses for each species, and extracts slope
}
\examples{
df = data.frame(species = c("A", "A", "A"), yr = c(1990, 1991, 1992), mass_mean = c(21, 23, 25))
regression_mass_slope(df)

}
