% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_mrt_p.R
\name{regression_mrt_p}
\alias{regression_mrt_p}
\title{Calculation of each species' temperature-mass relationship p-value and slope}
\usage{
regression_mrt_p(mass_df)
}
\arguments{
\item{mass_df}{dataframe containing all annual masses and temperatures for all species}
}
\value{
dataframe with species, slope, and p-value columns
}
\description{
Takes annual masses and corresponding temperatures for all species, runs
linear regression on relationship for each species, and extracts p-value
and slope
}
\examples{
df = data.frame(species = c("A", "A", "A"), yr = c(1990, 1991, 1992), avg_temp = c(10, 12, 14), mass_mean = c(21, 23, 25))
regression_mrt_p(df)

}
