% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_mrt_r2.R
\name{regression_mrt_r2}
\alias{regression_mrt_r2}
\title{Calculation of each species' temperature-mass relationship r^2}
\usage{
regression_mrt_r2(mass_df)
}
\arguments{
\item{mass_df}{dataframe containing all annual masses and temperatures for all species}
}
\value{
dataframe with species and r^2 columns
}
\description{
Takes annual masses and corresponding temperatures for all species, runs
linear regression on relationship for each species, and extracts r^2
}
\examples{
df = data.frame(species = c("A", "A", "A"), yr = c(1990, 1991, 1992), avg_temp = c(14, 12, 10), mass_mean = c(21, 23, 25))
regression_mrt_r2(df)

}
