% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_mass.R
\name{threshold_mass}
\alias{threshold_mass}
\title{Generate thresholded mass dataset}
\usage{
threshold_mass(full_mass_df, min_inds = 15, min_yrs = 5)
}
\arguments{
\item{full_mass_df}{entire mass dataset}

\item{min_inds}{desired minimum number of individuals}

\item{min_yrs}{desired minimum number of years}
}
\value{
thresholded mass dataset with species, inds, yr, site, avg_temp,
and num_yrs columns
}
\description{
Subset mass dataset to include only those species with a chosen minimum of
individuals and individuals from a chosen minimum of different years
}
\examples{
threshold_mass(threshold_mass_data)

}
