% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{threshold_mass_data}
\alias{threshold_mass_data}
\title{Example dataset for testing mass data thresholding function}
\format{A dataframe with xx rows and xx variables: 
\describe{
\item{col}{description}
}}
\usage{
threshold_mass_data
}
\description{
This is a made up set of data for three species at two sites to ensure the
threshold_mass function is working properly.
}
\keyword{datasets}
